A more const correct and consice Raymath.h
Swapping around returns, object creations, adding const to most everything.
There are some changes and objects removed that were unecessary
temporaries.

The only thing not comitted to is using restrict in the few places where
there are multiple pointers, which could be a slight improvement,
granted that it actually *should* be illegal to overlap memory regions
for those functions.